+++ src/Scripts/MusicPlayerLogic.lua
@@ -1,139 +1,92 @@
--- Checks of any of the songs have been pinned/favorited
-local function IsAnySongFavorited()
-	local favoritedTracks = {}
-	for _, songName in ipairs(game.ScreenData.MusicPlayer.Songs) do
-		if (songName ~= "Song_RandomSong" and songName ~= "Song_RandomSongFavorites") and game.GameState.WorldUpgrades[songName] and game.HasStoreItemPin(songName) then
-			table.insert(favoritedTracks, game.WorldUpgradeData[songName])
-		end
+function game.ModsNikkelMMusicMakerRandomizerShuffleFavorites(screen, button)
+	if not game.GameState.ModsNikkelMMusicMakerRandomizerShufflingFromFavorites then
+		game.GameState.MusicPlayerPlaylist = nil
+		game.GameState.ModsNikkelMMusicMakerRandomizerShufflingFromFavorites = true
 	end
-	return #favoritedTracks > 0
+	game.MusicPlayerShuffle(screen, button)
 end
 
-modutil.mod.Path.Wrap("SelectMusicPlayerItem", function(base, screen, button)
-	local components = button.Screen.Components
-	-- This will be set to the new playing song if the selection plays a random song
-	game.GameState.MusicMakerRandomizerFriendlyPlayingSong = "Nothing.."
-	game.CurrentRun.MusicMakerRandomizerTrackName = nil
-	game.CurrentRun.MusicMakerRandomizerTrackType = nil
+modutil.mod.Path.Wrap("MusicPlayerShuffle", function(base, screen, button)
+	-- Reset the flag if the function was not called through the favorite randomizer button
+	if not button.IsFavoriteShuffleButton then
+		game.GameState.MusicPlayerPlaylist = nil
+		game.GameState.ModsNikkelMMusicMakerRandomizerShufflingFromFavorites = false
+	end
+	base(screen, button)
+end)
 
-	-- Do not allow playing the favorites song if there are no favorited songs (still allow pausing it)
-	if game.GameState.MusicPlayerSongName ~= "Song_RandomSongFavorites" and button.Data.Name == "Song_RandomSongFavorites" and not IsAnySongFavorited() then
-		-- Display the default description for the favorites song 
-		game.ModifyTextBox(
-			{
-				Id = components.InfoBoxDescription.Id,
-				Text = "Song_RandomSongFavorites",
-				UseDescription = true
-			}
-		)
+-- Make the randomize button visible if applicable, and hide it otherwise
+modutil.mod.Path.Wrap("MusicPlayerDisplayItems", function(base, screen)
+	if game.GameState.WorldUpgrades.WorldUpgradeMusicPlayerShuffle then
+		SetAlpha({ Id = screen.Components.ShuffleButtonFavorites.Id, Fraction = 1.0, Duration = 0.2 })
+		screen.Components.ShuffleButtonFavorites.Visible = true
+	else
+		SetAlpha({ Id = screen.Components.ShuffleButtonFavorites.Id, Fraction = 0.0 })
+	end
+	base(screen)
+end)
 
-		-- Modify the voicelines that can play, excluding the ones that refer to not having enough resources
-		button.Data.CannotAffordVoiceLines =
-		{
-			BreakIfPlayed = true,
-			RandomRemaining = true,
-			PreLineWait = 0.15,
-			Cooldowns =
-			{
-				{ Name = "MelinoeAnyQuipSpeech" },
-				{ Name = "MelinoeResourceInteractionSpeech", Time = 6 },
-			},
-			{ Cue = "/VO/Melinoe_0386", Text = "I can't." },
-			{ Cue = "/VO/Melinoe_0390", Text = "No use." },
-			{ Cue = "/VO/Melinoe_0222", Text = "Can't." },
-			{ Cue = "/VO/Melinoe_1854", Text = "Afraid not..." },
-			{ Cue = "/VO/Melinoe_1855", Text = "Denied." },
-			{ Cue = "/VO/Melinoe_1856", Text = "Denied..." },
-		}
-		-- Animate the button as if an item could not be afforded/bought
-		game.ScreenCantAffordPresentation(screen, button)
+-- Add the pin component to all buttons (except the random song, which cannot be favorited), regardless of if purchased or not
+-- This is for the icon that shows on the right, not the button prompt at the bottom of the screen
+modutil.mod.Path.Wrap("MusicPlayerUpdateVisibility", function(base, screen, args)
+	base(screen, args)
 
+	if not game.GameState.WorldUpgradesAdded.WorldUpgradeMusicPlayerShuffle then
 		return
 	end
 
-	base(screen, button)
-
-	-- If the selected song was a random song, update the description to display the currently playing song's name, or "Nothing..." if the player paused the random song
-	if (game.GameState.MusicPlayerSongName == "Song_RandomSong" or game.GameState.MusicPlayerSongName == "Song_RandomSongFavorites" or game.GameState.MusicPlayerSongName == nil)
-			and (button.Data.Name == "Song_RandomSong" or button.Data.Name == "Song_RandomSongFavorites") then
-
-		local songItemIdentifier = "Song_RandomSong"
-		if button.Data.Name == "Song_RandomSong" then
-			songItemIdentifier = "Song_RandomSong_PlayingInfo"
-		elseif game.GameState.MusicPlayerSongName == nil then
-			songItemIdentifier = "Song_RandomSongFavorites"
-		else
-			songItemIdentifier = "Song_RandomSongFavorites_PlayingInfo"
-		end
-
-		game.ModifyTextBox(
-			{
-				Id = components.InfoBoxDescription.Id,
-				Text = songItemIdentifier,
-				UseDescription = true,
-				LuaKey = "TempTextData",
-				LuaValue = { PlayingSongFriendlyName = game.GameState.MusicMakerRandomizerFriendlyPlayingSong }
-			}
-		)
-
-		-- There is a chance for Melinoe to comment on playing a random song
-		if (game.GameState.MusicPlayerSongName == "Song_RandomSong" or game.GameState.MusicPlayerSongName == "Song_RandomSongFavorites") and game.RandomInt(1, 7) > 5 then
-			game.PlaySpeechCue("/VO/Melinoe_2356", nil, nil, "Interrupt", false)
-		end
-	end
-end)
-
--- Add the pin component to all buttons (except the random songs, they cannot be favorited), regardless of if purchased or not
--- This is for the icon that shows on the right, not the button prompt at the bottom of the screen
-modutil.mod.Path.Wrap("MusicPlayerDisplayItems", function(base, screen)
-	base(screen)
-
-	-- Reset and repeat the loop from the base function. The order will be the same
-	screen.NumItems = 0
 	local components = screen.Components
+	local firstIndex = screen.ScrollOffset + 1
+	local lastIndex = math.min(screen.NumItems, screen.ScrollOffset + screen.ItemsPerPage)
 
-	for i, songName in ipairs(screen.Songs) do
-		local songData = game.WorldUpgradeData[songName]
-
-		-- If the song has not been discovered yet, skip it
-		if songData.GameStateRequirements == nil or game.IsGameStateEligible(songData, songData.GameStateRequirements) then
-			screen.NumItems = screen.NumItems + 1
-			local pinButtonKey = "PinIcon" .. screen.NumItems
-
-			-- If the song already has a pin icon, skip it
-			if components[pinButtonKey] ~= nil or songName == "Song_RandomSong" then
-				goto continue
-			end
-
-			local purchaseButtonKey = "PurchaseButton" .. screen.NumItems
-			local button = components[purchaseButtonKey]
-
-			-- Create a "Pin" component for the button, which will be used for the favorite icon
-			components[pinButtonKey] = game.CreateScreenComponent({
-				Name = "BlankObstacle",
-				Group = screen.ComponentData.DefaultGroup,
-				Alpha = 0.0,
-				Scale = screen.PurchaseButtonScaleY
-			})
-
-			Attach({
-				Id = components[pinButtonKey].Id,
-				DestinationId = components[purchaseButtonKey].Id,
-				OffsetX = screen.PinOffsetX,
-				OffsetY = game.UIData.PinIconListOffsetY * screen.PurchaseButtonScaleY
-			})
-
-			components[purchaseButtonKey].PinButtonId = components[pinButtonKey].Id
-
-			-- If the song has already been favorited, add the icon
-			if game.HasStoreItemPin(button.Data.Name) then
-				components[purchaseButtonKey].IsPinned = true
-				SetAnimation({
-					Name = "ModsNikkelMMusicMakerRandomizerFavorite",
-					DestinationId = components[purchaseButtonKey].PinButtonId
-				})
-			end
+	for itemIndex = firstIndex, lastIndex, 1 do
+		local itemData = screen.AvailableItems[itemIndex]
+		local displayName = itemData.Name
+		local purchased = GameState.WorldUpgrades[displayName]
+		if not purchased then
+			goto continue
 		end
+
+		local pinButtonKey = displayName .. "PinIcon"
+		-- If the song already has a pin icon, destroy and rebuild it
+		if components[pinButtonKey] ~= nil then
+			Destroy({ Id = components[pinButtonKey].Id })
+		end
+
+		local buttonKey = displayName .. "Button"
+		local button = components[buttonKey]
+
+		-- Create a "Pin" component for the button, which will be used for the favorite icon
+		local pinIcon = game.CreateScreenComponent({
+			Name = "BlankObstacle",
+			Group = screen.ComponentData.DefaultGroup,
+			Alpha = 0.0,
+			Scale = screen.PinScale,
+		}) or {}
+		components[pinButtonKey] = pinIcon
+		table.insert(button.AssociatedIds, pinIcon.Id)
+
+		Attach({
+			Id = pinIcon.Id,
+			DestinationId = button.Id,
+			OffsetX = screen.PinOffsetX,
+			OffsetY = game.UIData.PinIconListOffsetY + screen.PinOffsetY
+		})
+
+		button.PinButtonId = pinIcon.Id
+
+		-- If the song has already been favorited, add the icon
+		if game.Contains(game.GameState.ModsNikkelMMusicMakerRandomizerFavoritedTracks, displayName) then
+			button.IsPinned = true
+			SetAnimation({
+				Name = "ModsNikkelMMusicMakerRandomizerFavorite",
+				DestinationId = button.PinButtonId
+			})
+		end
+		SetAlpha({ Id = button.Id, Fraction = 1.0, Duration = 0.1 })
+		SetAlpha({ Ids = button.AssociatedIds, Fraction = 1.0, Duration = 0.1 })
+		table.insert(screen.ButtonIds, button.Id)
+		screen.ButtonIds = game.CombineTables(screen.ButtonIds, button.AssociatedIds)
 
 		::continue::
 	end