+++ src/config.lua
@@ -1,12 +1,9 @@
 local config = {
-  version = 0;
   Enabled = true;
-  AllSongs = false;
 }
 
 local configDesc = {
   Enabled = "Whether the mod is enabled or not.";
-  AllSongs = "Whether the Music Maker should be able to shuffle from all songs in the game, even those not yet discovered or unlocked yet (SPOILER WARNING)."
 }
 
 return config, configDesc